backend_api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
google-generativeai==0.8.4
python-dotenv==1.1.0
requests==2.32.3
pillow==11.1.0
